backend/requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain_openai
langchain_google_genai
langchain_groq
pydantic
faiss-cpu
python-dotenv
mcp
numpy
scikit-learn
requests
motor
python-jose
httpx
cryptography
duckduckgo-search
beautifulsoup4
praw
tweepy
